moved orthogonal and curtain
